Separate bitmask implementation from old implementation
